notebooks/requirements.txt
ipykernel==6.30.1
google-cloud-secret-manager==2.24.0
matplotlib==3.10.6
nbclient==0.10.2
nbformat==5.10.4
numpy==2.3.3
pandas==2.3.2
pandas-gbq==0.29.2
plotly==6.3.0
pyYAML==6.0.2
tqdm==4.67.1